refactor: domain -> domain.ticket
