+++ src/main/kotlin/domain/LottoPrizeCalculator.kt
@@ -1,6 +1,7 @@
 package org.example.domain
 
 import org.example.LottoCheckRank
+import org.example.domain.ticket.WinningLotto
 
 class LottoPrizeCalculator(
     private val userLottoTickets: LottoTickets,
